Rename: dom-api 파일들 하나로 병합
