--- js/dom-api/getElementByClassName.js
@@ -1,11 +1,0 @@
-export const getElementByClassName = (node, target) => {
-  let result = null;
-  if (!node) return null;
-  for (let i = 0; i < node.children.length; i++) {
-    if ([...node.children[i].classList].includes(target)) {
-      return node.children[i];
-    }
-    result = getElementByClassName(node.children[i], target);
-    if (result) return result;
-  }
-};
